chore: refactor analytics backend dialog into separate component
